requirements.txt
boto3>=1.26.0
pandas>=1.5.0
botocore>=1.29.0
black>=23.0.0
streamlit>=1.30.0
altair>=5.0.0
gspread>=5.0.0
google-auth-oauthlib>=0.7.0 